# fused adam(w): Reduce register usage