@crcrpar 2024-01-26
There are I think two major differences between pytorch fused adam(w) and apex fused adam(w), so I'd like to split this pull request into two
- indexing: pytorch uses 64 bits by default while apex 32
  -  #118316
- register usage: as in the table above, 80 vs 48.